requirements.txt
asgiref==3.8.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
Django==5.1.7
djangorestframework==3.15.2
dotenv==0.9.9
drf-yasg==1.21.8
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
isort==6.0.0
msgpack==1.1.0
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
